Table Products

+++ database/migrations/2022_04_06_235308_create_products_table.php
@@ -17,7 +17,7 @@ class CreateProductsTable extends Migration
             $table->id();
             $table->string('name');
             $table->string('description');
-            $table->unsignedBigInteger('price');
+            $table->unsignedInteger('price');
             $table->timestamps();
         });
     }
